small .NET side refactor
