+++ src/Designer/backend/src/Designer/Infrastructure/ServiceRegistration.cs
@@ -105,6 +105,7 @@ public static class ServiceRegistration
         services.AddTransient<IOrgContentService, OrgContentService>();
         services.AddTransient<IEnvironmentsService, EnvironmentsService>();
         services.AddSingleton<IStudioctlInstallScriptService, StudioctlInstallScriptService>();
+        services.AddScoped<StudioctlAuthService>();
         services.AddHttpClient<IOrgService, OrgService>();
         services.AddHttpClient<ImageClient>();
         services.AddTransient<IAppDevelopmentService, AppDevelopmentService>();
